requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiohttp-swagger==1.0.16
aiosignal==1.3.1
arrow==1.2.3
async-lru==2.0.2
async-timeout==4.0.2
attrs==23.1.0
backports.ssl==0.0.9
bcrypt==4.0.1
binaryornot==0.4.4
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
cookiecutter==2.1.1
cryptography==40.0.2
dill==0.3.6
distro==1.8.0
frozenlist==1.3.3
genie==23.5
genie.libs.clean==23.5
genie.libs.conf==23.5
genie.libs.filetransferutils==23.5
genie.libs.health==23.5
genie.libs.ops==23.5
genie.libs.parser==23.5
genie.libs.robot==23.5
genie.libs.sdk==23.5
genie.telemetry==23.5
genie.trafficgen==23.5
gitdb==4.0.10
GitPython==3.1.31
grpcio==1.54.2
idna==3.4
IxNetwork==9.0.1915.16
ixnetwork-restpy==1.1.9
Jinja2==3.1.2
jinja2-time==0.2.0
jsonpickle==3.0.1
junit-xml==1.9
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
ncclient==0.6.13
netaddr==0.8.0
packaging==23.1
paramiko==3.2.0
pathspec==0.11.1
prettytable==3.7.0
protobuf==3.20.3
psutil==5.9.5
pyats==23.5
pyats.aereport==23.5
pyats.aetest==23.5
pyats.async==23.5
pyats.connections==23.5
pyats.contrib==23.5
pyats.datastructures==23.5
pyats.easypy==23.5
pyats.kleenex==23.5
pyats.log==23.5
pyats.reporter==23.5
pyats.results==23.5
pyats.robot==23.5
pyats.tcl==23.5
pyats.topology==23.5
pyats.utils==23.5
pycparser==2.21
pyftpdlib==1.5.7
PyNaCl==1.5.0
pyOpenSSL==23.1.1
python-dateutil==2.8.2
python-engineio==3.14.2
python-slugify==8.0.1
python-socketio==4.6.1
PyYAML==6.0
requests==2.31.0
requests-toolbelt==1.0.0
robotframework==6.0.2
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
six==1.16.0
smmap==5.0.0
text-unidecode==1.3
tftpy==0.8.0
tqdm==4.65.0
typing_extensions==4.6.2
unicon==23.5
unicon.plugins==23.5
urllib3==2.0.2
wcwidth==0.2.6
websocket-client==1.5.2
xlrd==1.2.0
XlsxWriter==3.1.2
xlwt==1.3.0
xmltodict==0.13.0
yamllint==1.32.0
yang.connector==23.5
yarl==1.9.2
